change the relocations behavior for kaslr on x86_64
On x86_64, in old kaslr implementaion only physical address of kernel
loading is randomized. Then calculate the delta value of physical address
where vmlinux was linked to load and where it is finally loaded. If
the delta is not equal to 0, namely there's a new physical address where
kernel is actually decompressed, relocation handling need be done. Then
delta is added to offset of kernel symbol relocation, this makes the
address of kernel text mapping move delta long.

Now the old behavior is changed. We randomize both the physical address
where kernel is decompressed and the virtual address where kernel text
is mapped. And here relocation handling only depends on virtual address
randomization. Means if and only if virtual address of kernel text mapping
is randomized to a different value, we add the delta between
LOAD_PHYSICAL_ADDR and CONFIG_RANDOMIZE_BASE_MAX_OFFSET to the offset
of kernel relocs.

Signed-off-by: Baoquan He <bhe@redhat.com>